+++ src/main/java/dk/sdu/se_f22/pim2/ProductIndexInfrastructure.java
@@ -1,10 +1,16 @@
 package dk.sdu.se_f22.pim2;
 
+import dk.sdu.se_f22.pim2.domain.ProductInfIndex;
+import dk.sdu.se_f22.pim2.domain.ProductInfIndexImpl;
+import dk.sdu.se_f22.pim2.domain.ProductInfSearch;
+import dk.sdu.se_f22.pim2.domain.ProductInfSearchImpl;
+
 public class ProductIndexInfrastructure {
-    private ProductIndexInfrastructure() {
+    private static final ProductIndexInfrastructure INSTANCE = new ProductIndexInfrastructure();
+
+	private ProductIndexInfrastructure() {
     }
 
-    private static final ProductIndexInfrastructure INSTANCE = new ProductIndexInfrastructure();
 
     public ProductInfSearch getProductSearch() {
         return new ProductInfSearchImpl();
