+++ src/main/java/dk/sdu/se_f22/pim2/domain/ProductInfSearchImpl.java
@@ -1,4 +1,6 @@
-package dk.sdu.se_f22.pim2;
+package dk.sdu.se_f22.pim2.domain;
+
+import dk.sdu.se_f22.pim2.Product;
 
 import java.util.List;
